requirements.txt
Flask
SQLAlchemy
requests
gunicorn
pytest
flask_bcrypt
flask_login
flask_cors
mysql-connector-python
mysql
api
python-dotenv
pyodbc
pandas